try fix ResourceManager test
