+++ ingest/src/main/java/com/microsoft/azure/kusto/ingest/ResourceManager.java
@@ -198,6 +198,7 @@ class ResourceManager implements Closeable {
                         addIngestionResource(resourceTypeName, storageUrl);
                     }
                 }
+                log.info("Refreshing Ingestion Resources Finised");
             } catch (DataServiceException e) {
                 throw new IngestionServiceException(e.getIngestionSource(), "Error refreshing IngestionResources", e);
             } catch (DataClientException e) {
